backend/requirements.txt
fastapi==0.135.1
uvicorn==0.41.0
python-multipart==0.0.6
pillow
numpy
tensorflow==2.19.0
tensorflow_hub==0.16.1
keras==3.10.0
opencv-python-headless==4.8.1.78
python-dotenv==1.0.0